+++ CHANGELOG.md
@@ -70,7 +70,7 @@ All notable changes to this project will be documented in this file.
     - Remove validator verification via gossip. This functionality is migrated to AccessPass.
 - Device controller
     - Implement user tunnel ACLs in device agent configuration
-    - Make MaxTunnelSlots configurable
+    - Replace MaxTunnelSlots constant with device.max_users from serviceability. Max users is now configurable on a per-device bases.
     - Add "mpls icmp ttl-exceeded tunneling" config statement so intermediate hops in the doublezero network respond to traceroutes.
     - Set protocol timers for ibgp and isis to improve to speed up network re-convergence
     - Add TLS support to gRPC server